+++ src/main/java/study/codereview/book/exception/exceptions/CostDivideException.java
@@ -1,4 +1,4 @@
-package study.codereview.book.domain.exception.exceptions;
+package study.codereview.book.exception.exceptions;
 
 public class CostDivideException extends RuntimeException {
