+++ src/main/java/study/codereview/book/exception/exceptions/CostRangeException.java
@@ -1,4 +1,4 @@
-package study.codereview.book.domain.exception.exceptions;
+package study.codereview.book.exception.exceptions;
 
 public class CostRangeException extends RuntimeException {
